TST: move misplaced tests
